Resolve menu issues when waking from sleep
The menu would be instantly closed by updateMenu() until the next refresh. This was due to the NSTimer's fireDate being in the past, not aligning with the true refresh, leading to a negative time remaining causing the menu to close expecting a refresh. Solution is to check for wakes and make sure the timer is in sync.